Store localization grids using single precision instead of double precision for reducing memory consumption. We presume that single precision is sufficient because the values stored in the `offsets` array are differences, not absolute coordinates. The absolute coordinates will be computed by adding `double` values to those offsets.

+++ core/sis-referencing/src/test/java/org/apache/sis/referencing/operation/builder/ResidualGridTest.java
@@ -47,7 +47,7 @@ public final strictfp class ResidualGridTest extends TestCase {
      * The grid has no "source to grid" or "grid to CRS" transformations.
      */
     public ResidualGridTest() {
-        grid = new ResidualGrid(MathTransforms.identity(2), MathTransforms.identity(2), 3, 4, new double[] {
+        grid = new ResidualGrid(MathTransforms.identity(2), MathTransforms.identity(2), 3, 4, new float[] {
                 0,2  ,  1,2  ,  2,1,
                 1,3  ,  2,2  ,  1,1,
                 0,4  ,  2,3  ,  3,2,
